backend/pyproject.toml
[project]
name = "timetracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2,<5.0",
    "django-cors-headers>=4.3",
    "django-filter>=25.1",
    "djangorestframework>=3.15",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
]
